refactor: New structure for the columns
